requirements.txt
Django>=5.0,<5.1
psycopg2-binary>=2.9.9
pytesseract>=0.3.10
pdfplumber>=0.10.3
Pillow>=10.2.0
gunicorn>=21.2.0
python-dotenv>=1.0.1
django-environ>=0.11.2